Move UsageStore into media-api
